Remove this odd watch test
